Fixing config direcotry function to match new specs
