+++ pathing/config_file_unix.go
@@ -2,7 +2,43 @@
 
 package pathing
 
+import (
+	"log"
+	"os"
+	"path/filepath"
+)
+
 const (
 	defaultConfigFile = ".packerconfig"
 	defaultConfigDir  = ".packer.d"
 )
+
+func configDir() (path string, err error) {
+	var dir string
+	homedir := os.Getenv("HOME")
+	if homedir == "" {
+		return "", err
+	}
+	if cd := os.Getenv("PACKER_CONFIG_DIR"); cd != "" {
+		log.Printf("Detected config directory from env var: %s", cd)
+		dir = filepath.Join(cd, defaultConfigDir)
+	} else if hasDefaultConfigFileLocation(homedir) {
+		dir = filepath.Join(homedir, defaultConfigDir)
+		log.Printf("Old default config directory found: %s", dir)
+	} else if xdgConfigHome := os.Getenv("XDG_CONFIG_HOME"); xdgConfigHome != "" {
+		log.Printf("Detected xdg config directory from env var: %s", xdgConfigHome)
+		dir = xdgConfigHome
+	} else {
+		dir = filepath.Join(homedir, ".config")
+	}
+
+	return dir, nil
+}
+
+func hasDefaultConfigFileLocation(homedir string) bool {
+	_, err := os.Stat(filepath.Join(homedir, defaultConfigDir))
+	if err == nil {
+		return true
+	}
+	return false
+}
